requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
openai==1.3.7
Werkzeug==2.3.7
